[major change] do not treat the name the app uses as the filename
With this change, it's possible for the app to refer to, say /js/main.js
without there being a /js directory anywhere.

In other words: where the file is read from, and where the hash-prefixed
file is written to is now completely independent from the path of the
file that the app uses.